requirements.txt
crewai==0.51.1
groq==0.9.0
langchain_groq==0.1.9
pydantic==1.10.12
python-dotenv==1.0.1
strictyaml==1.7.3